add separate api and private/unit test executables
Related: https://github.com/elucideye/acf/issues/94

We need a separate executables for:
(1) the public (possibly shared) ACF api
(2) the private support class/functions used internally in ACF that require focused testing

* add SIMD channel conversion code to ACF (imgrated from drishti) and to private unit test
* move optimized separable triangle filter shader test to private unit test
* extract ogles_gpgpu::ACF::getImage() texture read code to acf/transfer.{h,cpp} as stand-alone functions for easier reuse (without require compilatino of GPUACF.{h,cpp}
* collect private unit test sources with sugar in ACF_TEST_SRCS variable and build into the acf unit test executable
