+++ src/sugar.cmake
@@ -1,0 +1,1 @@
+sugar_include(lib)
