+++ src/test/CMakeLists.txt
@@ -1,24 +1,28 @@
 # Copyright (c) 2015-2017, Ruslan Baratov, David Hirvonen
 # All rights reserved.
 
-set(test_name ACFTest)
-set(test_app test-acf)
+################
+### api test ### : STATIC or SHARED w/ hidden symbols
+################
 
-set(acf_ut_srcs test-acf.cpp)
+set(test_api_name ACFTest_api)
+set(test_api_app test-acf-api)
 
-add_executable(${test_app} ${acf_ut_srcs})
-set_property(TARGET ${test_app} PROPERTY FOLDER "app/tests")
-target_link_libraries(${test_app} PUBLIC acf::acf acf_common acf_shaders GTest::gtest ${OpenCV_LIBS})
+set(acf_ut_srcs test-acf-api.cpp)
+
+add_executable(${test_api_app} ${acf_ut_srcs})
+set_property(TARGET ${test_api_app} PROPERTY FOLDER "app/tests")
+target_link_libraries(${test_api_app} PUBLIC acf::acf acf_common GTest::gtest ${OpenCV_LIBS})
 
 if (ACF_BUILD_OGLES_GPGPU AND TARGET aglet::aglet)
-  target_link_libraries(${test_app} PUBLIC aglet::aglet)
-  target_compile_definitions(${test_app} PUBLIC ACF_DO_GPU=1)
+  target_link_libraries(${test_api_app} PUBLIC aglet::aglet)
+  target_compile_definitions(${test_api_app} PUBLIC ACF_DO_GPU=1)
 endif()
 
 # GTest + CTest
 gauze_add_test(
-  NAME ${test_name}
-  COMMAND ${test_app}
+  NAME ${test_api_name}
+  COMMAND ${test_api_app}
   "$<GAUZE_RESOURCE_FILE:${DRISHTI_FACES_FACE_IMAGE}>"
   "$<GAUZE_RESOURCE_FILE:${DRISHTI_FACES_FACE_IMAGE}>"
   "$<GAUZE_RESOURCE_FILE:${DRISHTI_ASSETS_FACE_DETECTOR}>"
@@ -29,9 +33,33 @@ gauze_add_test(
 
 if(WIN32 OR CYGWIN)
   set_property(
-    TEST ${test_name}
+    TEST ${test_api_name}
     PROPERTY
     ENVIRONMENT
     "PATH=$<TARGET_FILE_DIR:acf>;$ENV{PATH}"
    )
 endif()
+
+#################
+### unit test ### : this will always use static linking
+#################
+
+set(test_unit_name ACFTest_unit)
+set(test_unit_app test-acf-unit)
+
+set(acf_ut_srcs test-acf-unit.cpp ${ACF_TEST_SRCS})
+
+add_executable(${test_unit_app} ${acf_ut_srcs})
+set_property(TARGET ${test_unit_app} PROPERTY FOLDER "app/tests")
+target_link_libraries(${test_unit_app} PUBLIC acf_headers acf_common GTest::gtest ${OpenCV_LIBS})
+
+if (ACF_BUILD_OGLES_GPGPU AND TARGET aglet::aglet)
+  target_link_libraries(${test_unit_app} PUBLIC aglet::aglet ogles_gpgpu::ogles_gpgpu)
+  target_compile_definitions(${test_unit_app} PUBLIC ACF_DO_GPU=1)
+endif()
+
+# GTest + CTest
+gauze_add_test(
+  NAME ${test_unit_name}
+  COMMAND ${test_unit_app}
+)
